move Equipment Sorting to Utility file, add querying categories
